+++ src/app/disease-form/common/hexHexDummy.ts
@@ -183,6 +183,15 @@ export class HexHexDummy {
               valueString: 'Wichtiger Hinweis zum Status der betroffenen Person.',
             },
           },
+          ...(environment.featureFlags?.FEATURE_FLAG_REFERENCE_FIELD
+            ? {
+                notificationIdReference: {
+                  answer: {
+                    valueString: 'noReference',
+                  },
+                },
+              }
+            : {}),
           initialNotificationId: '',
         },
       },
@@ -885,6 +894,15 @@ export class HexHexDummy {
               valueString: 'Wichtiger Hinweis zum Status der betroffenen Person.',
             },
           },
+          ...(environment.featureFlags?.FEATURE_FLAG_REFERENCE_FIELD
+            ? {
+                notificationIdReference: {
+                  answer: {
+                    valueString: 'noReference',
+                  },
+                },
+              }
+            : {}),
           initialNotificationId: '',
         },
       },